+++ cpp/tests/utils/TimeUtils.h
@@ -33,8 +33,8 @@ namespace TimeUtils
  */
 inline static std::uint64_t getCurrentMillisSinceEpoch()
 {
-    return std::chrono::duration_cast<std::chrono::milliseconds>(
-                   std::chrono::system_clock::now().time_since_epoch()).count();
+    return static_cast<std::uint64_t>(std::chrono::duration_cast<std::chrono::milliseconds>(
+                   std::chrono::system_clock::now().time_since_epoch()).count());
 }
 } // namespace TimeUtils
